LayoutTests/imported/w3c/web-platform-tests/scroll-animations/scroll-timelines/scroll-animation-effect-fill-modes.tentative-expected.txt


PASS Scroll based animation effect fill mode should return 'auto' for getTiming() and should return 'none' for getComputedTiming().
FAIL Applied effect value before start delay with fill: none assert_equals: animation effect applied property value expected 1 but got 0.34
FAIL Applied effect value at start delay with fill: none assert_equals: animation effect applied property value expected 0.3 but got 0.4
PASS Applied effect value at midpoint with fill: none
FAIL Applied effect value at effect end with fill: none assert_equals: animation effect applied property value expected 1 but got 0.6
FAIL Applied effect value after effect end with fill: none assert_equals: animation effect applied property value expected 1 but got 0.66
FAIL Applied effect value before start delay with fill: backwards assert_equals: animation effect applied property value expected 0.3 but got 0.34
FAIL Applied effect value at start delay with fill: backwards assert_equals: animation effect applied property value expected 0.3 but got 0.4
PASS Applied effect value at midpoint with fill: backwards
FAIL Applied effect value at effect end with fill: backwards assert_equals: animation effect applied property value expected 1 but got 0.6
FAIL Applied effect value after effect end with fill: backwards assert_equals: animation effect applied property value expected 1 but got 0.66
FAIL Applied effect value before timeline start with fill: forwards assert_equals: animation effect applied property value expected 1 but got 0.34
FAIL Applied effect value at timeline start with fill: forwards assert_equals: animation effect applied property value expected 0.3 but got 0.4
PASS Applied effect value in timeline range with fill: forwards
FAIL Applied effect value at timeline end with fill: forwards assert_equals: animation effect applied property value expected 0.7 but got 0.6
FAIL Applied effect value after timeline end with fill: forwards assert_equals: animation effect applied property value expected 0.7 but got 0.66
FAIL Applied effect value before timeline start with fill: both assert_equals: animation effect applied property value expected 0.3 but got 0.34
FAIL Applied effect value at timeline start with fill: both assert_equals: animation effect applied property value expected 0.3 but got 0.4
PASS Applied effect value in timeline range with fill: both
FAIL Applied effect value at timeline end with fill: both assert_equals: animation effect applied property value expected 0.7 but got 0.6
FAIL Applied effect value after timeline end with fill: both assert_equals: animation effect applied property value expected 0.7 but got 0.66
FAIL Applied effect value before timeline start with fill: auto assert_equals: animation effect applied property value expected 1 but got 0.34
FAIL Applied effect value at timeline start with fill: auto assert_equals: animation effect applied property value expected 0.3 but got 0.4
PASS Applied effect value in timeline range with fill: auto
FAIL Applied effect value at timeline end with fill: auto assert_equals: animation effect applied property value expected 1 but got 0.6
FAIL Applied effect value after timeline end with fill: auto assert_equals: animation effect applied property value expected 1 but got 0.66
